scripts/requirements.txt
mxnet-mkl==1.6.0
gluoncv==0.8.0